requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pypdf2>=3.0.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pytesseract>=0.3.10
easyocr>=1.7.0

# NLP and Text Processing
spacy>=3.7.0
nltk>=3.8.0
regex>=2023.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image Processing
matplotlib>=3.7.0
seaborn>=0.12.0
albumentations>=1.3.0

# Configuration and Utilities
PyYAML>=6.0
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU support
# torch-cuda>=2.0.0  # Uncomment if using CUDA 